+++ packages/editor-common/web/src/index.ts
@@ -9,8 +9,6 @@ export { default as FocusManager } from './Components/FocusManager';
 export { default as LinkPanelWrapper } from './Components/LinkComponents/LinkPanelWrapper';
 export { default as LinkButton } from './Components/LinkComponents/LinkButton';
 export { default as LinkModal } from './Components/LinkComponents/LinkModal';
-export { default as ToolbarButton } from './Components/ToolbarButton';
-export { default as InlineToolbarButton } from './Components/InlineToolbarButton';
 
 //Modals
 export { default as EditorModals } from './Modals/EditorModals';
@@ -91,7 +89,6 @@ export {
   FORMATTING_BUTTONS,
   INSERT_PLUGIN_BUTTONS,
   BUTTON_TYPES,
-  KEYS_CHARCODE,
 } from './consts';
 
 import './draftTypes';