+++ packages/ui-components/web/src/Icons/index.ts
@@ -28,5 +28,6 @@ export { default as CloseIcon } from './CloseIcon';
 export { default as MoreIcon } from './MoreIcon';
 export { default as ClearIcon } from './ClearIcon';
 export { default as SearchIcon } from './SearchIcon';
-export { default as DropdownArrowIcon } from './DropdownArrowIcon';
 export { default as CheckIcon } from './CheckIcon';
+export { default as DropdownArrowIcon } from './DropdownArrowIcon';
+export { default as ErrorIcon } from './ErrorIcon';